+++ src/main/java/cc/altius/FASP/model/report/StockStatusVerticalOutputRowMapper.java
@@ -29,28 +29,28 @@ public class StockStatusVerticalOutputRowMapper implements ResultSetExtractor<Li
             ssv.setDt(rs.getDate("TRANS_DATE"));
             int index = ssvList.indexOf(ssv);
             if (index == -1) {
-                ssv.setOpeningBalance(rs.getInt("FINAL_OPENING_BALANCE"));
+                ssv.setOpeningBalance(rs.getLong("FINAL_OPENING_BALANCE"));
                 ssv.setActualConsumption(rs.getBoolean("ACTUAL"));
                 if (rs.wasNull()) {
                     ssv.setActualConsumption(null);
                 }
-                ssv.setConsumptionQty(rs.getInt("CONSUMPTION_QTY"));
+                ssv.setConsumptionQty(rs.getLong("CONSUMPTION_QTY"));
                 if (rs.wasNull()) {
                     ssv.setConsumptionQty(null);
                 }
-                ssv.setShipmentQty(rs.getInt("SQTY"));
+                ssv.setShipmentQty(rs.getLong("SQTY"));
                 if (rs.wasNull()) {
                     ssv.setShipmentQty(null);
                 }
-                ssv.setAdjustment(rs.getInt("ADJUSTMENT"));
+                ssv.setAdjustment(rs.getLong("ADJUSTMENT"));
                 if (rs.wasNull()) {
                     ssv.setAdjustment(null);
                 }
-                ssv.setExpiredStock(rs.getInt("EXPIRED_STOCK"));
+                ssv.setExpiredStock(rs.getLong("EXPIRED_STOCK"));
                 if (rs.wasNull()) {
                     ssv.setExpiredStock(null);
                 }
-                ssv.setClosingBalance(rs.getInt("FINAL_CLOSING_BALANCE"));
+                ssv.setClosingBalance(rs.getLong("FINAL_CLOSING_BALANCE"));
                 if (rs.wasNull()) {
                     ssv.setClosingBalance(null);
                 }
